Allow settings imposter host in constructor
Allow setting the hostname when creating an imposter, rather than have it hardcoded to `localhost`. 

This is helpful when connecting to a running Mountebank instance in a docker-compose setup in combination with the MountebankServer class - the hostname needs to be the docker-compose service name.

We set the default to localhost for the standard use case.